compiler

+++ src/main/java/com/coperatecoding/secodeverseback/service/QuestionImgService.java
@@ -1,8 +1,11 @@
 package com.coperatecoding.secodeverseback.service;
 
+import com.coperatecoding.secodeverseback.domain.RoleType;
+import com.coperatecoding.secodeverseback.domain.User;
 import com.coperatecoding.secodeverseback.domain.question.Question;
 import com.coperatecoding.secodeverseback.domain.question.QuestionImage;
 import com.coperatecoding.secodeverseback.dto.QuestionImgDTO;
+import com.coperatecoding.secodeverseback.exception.ForbiddenException;
 import com.coperatecoding.secodeverseback.exception.NotFoundException;
 import com.coperatecoding.secodeverseback.repository.QuestionImageRepository;
 import com.coperatecoding.secodeverseback.repository.QuestionRepository;
@@ -73,9 +76,12 @@ public class QuestionImgService {
 
         return response;
     }
-    public void delete(Long imgPk) throws RuntimeException {
+    public void delete(User user, Long imgPk) throws RuntimeException {
         QuestionImage questionImage = questionImageRepository.findById(imgPk)
                 .orElseThrow(() -> new NotFoundException("해당하는 이미지가 존재하지 않음"));
+        if(!user.getRoleType().equals(RoleType.ADMIN) && questionImage.getUser().getPk() != user.getPk()) {
+            throw new ForbiddenException("작성자만 댓글을 삭제할 수 있습니다.");
+        }
 
         questionImageRepository.delete(questionImage);
     }
